requirements.txt
requests
python-Wappalyzer
beautifulsoup4
setuptools
